more and more cleanup
