+++ packages/studiocms/src/routes/api/auth/[provider]/_effects/index.ts
@@ -1,82 +1,11 @@
-import { type AuthEnvCheckResponse, authEnvCheck } from 'studiocms:auth/utils/authEnvCheck';
-import { authConfig } from 'studiocms:config';
+import type { AuthEnvCheckResponse } from 'studiocms:auth/utils/authEnvCheck';
 import type { APIContext } from 'astro';
-import { Context, Data, Effect, Layer, genLogger } from '../../../../effect.js';
-import { Auth0OAuthAPI } from './_effects/auth0.js';
-import { DiscordOAuthAPI } from './_effects/discord.js';
-import { GitHubOAuthAPI } from './_effects/github.js';
-import { GoogleOAuthAPI } from './_effects/google.js';
-
-/**
- * Enum representing supported authentication providers.
- *
- * @remarks
- * Used to specify the provider for authentication flows.
- *
- * @enum {string}
- * @property {Provider.GOOGLE}  Google authentication provider.
- * @property {Provider.GITHUB}  GitHub authentication provider.
- * @property {Provider.DISCORD} Discord authentication provider.
- * @property {Provider.AUTH0}   Auth0 authentication provider.
- */
-export enum Provider {
-	GOOGLE = 'google',
-	GITHUB = 'github',
-	DISCORD = 'discord',
-	AUTH0 = 'auth0',
-}
-
-/**
- * Creates a standardized HTTP response for authentication provider errors.
- *
- * @param error - The error message to include in the response body.
- * @param status - The HTTP status code for the response.
- * @returns A `Response` object with a JSON body containing the error message and the specified status code.
- */
-export const ProviderResponse = (
-	error: string,
-	status: number
-): Effect.Effect<Response, never, never> =>
-	Effect.succeed(new Response(JSON.stringify({ error }), { status }));
-
-/**
- * Represents an error related to authentication environment configuration.
- *
- * This error is tagged as 'AuthEnvError' and includes a message describing the issue.
- *
- * @extends Data.TaggedError
- * @template { message: string } - The payload containing the error message.
- */
-export class AuthEnvError extends Data.TaggedError('AuthEnvError')<{ message: string }> {}
-
-/**
- * Checks the authentication environment configuration for all providers.
- *
- * This function wraps the `authEnvCheck` call in an `Effect.tryPromise`, handling any errors
- * by returning an `AuthEnvError` with a descriptive message.
- *
- * @returns {Effect<unknown, AuthEnvError, void>} An effect that resolves if the environment is valid,
- * or fails with an `AuthEnvError` if the check fails.
- */
-export const authEnvChecker = (): Effect.Effect<AuthEnvCheckResponse, AuthEnvError, never> =>
-	Effect.tryPromise({
-		try: () => authEnvCheck(authConfig.providers),
-		catch: (cause) =>
-			new AuthEnvError({ message: `Authentication environment check failed: ${cause}` }),
-	});
-
-export class AuthEnvCheck extends Context.Tag('AuthEnvCheck')<
-	AuthEnvCheck,
-	AuthEnvCheckResponse
->() {
-	static make = (response: AuthEnvCheckResponse) => Layer.succeed(this, this.of(response));
-	static Provide = (response: AuthEnvCheckResponse) => Effect.provide(this.make(response));
-}
-
-export class ValidateAuthCodeError extends Data.TaggedError('ValidateAuthCodeError')<{
-	message: string;
-	provider: string;
-}> {}
+import { Effect, genLogger } from '../../../../../effect.js';
+import { AuthEnvCheck, Provider, ProviderResponse, authEnvChecker } from './_shared.js';
+import { Auth0OAuthAPI } from './auth0.js';
+import { DiscordOAuthAPI } from './discord.js';
+import { GitHubOAuthAPI } from './github.js';
+import { GoogleOAuthAPI } from './google.js';
 
 export class OAuthAPIEffect extends Effect.Service<OAuthAPIEffect>()('OAuthAPIEffect', {
 	dependencies: [
@@ -175,4 +104,6 @@ export class OAuthAPIEffect extends Effect.Service<OAuthAPIEffect>()('OAuthAPIEf
 }) {
 	static A = Effect.provide(OAuthAPIEffect.Default);
 	static B = (response: AuthEnvCheckResponse) => AuthEnvCheck.Provide(response);
+
+    static envChecker = async () => Effect.runPromise(authEnvChecker());
 }
